Unify file naming
