+++ roxctl/netpol/connectivity/diff/diff.go
@@ -1,107 +1,90 @@
+// Package diff provides primitives for command 'roxctl connectivity diff'
 package diff
 
 import (
 	"os"
+	"path/filepath"
 
-	npguard "github.com/np-guard/netpol-analyzer/pkg/netpol/diff"
+	npgdiff "github.com/np-guard/netpol-analyzer/pkg/netpol/diff"
 	"github.com/pkg/errors"
-	"github.com/spf13/cobra"
 	"github.com/stackrox/rox/pkg/errox"
-	"github.com/stackrox/rox/roxctl/common"
-	"github.com/stackrox/rox/roxctl/common/environment"
 	"github.com/stackrox/rox/roxctl/common/npg"
 	"github.com/stackrox/rox/roxctl/netpol/resources"
+	"k8s.io/cli-runtime/pkg/resource"
 )
 
-type diffNetpolCommand struct {
-	// Properties that are bound to cobra flags.
-	stopOnFirstError      bool
-	treatWarningsAsErrors bool
-	inputFolderPath1      string
-	inputFolderPath2      string
-	outputFilePath        string
-	removeOutputPath      bool
-	outputToFile          bool
-	outputFormat          string
+const (
+	defaultOutputFileNamePrefix = "connectivity_diff."
+	defaultOutputFormat         = "txt"
+)
 
-	// injected or constructed values
-	env environment.Environment
+type diffAnalyzer interface {
+	ConnDiffFromResourceInfos(infos1, infos2 []*resource.Info) (npgdiff.ConnectivityDiff, error)
+	ConnectivityDiffToString(connectivityDiff npgdiff.ConnectivityDiff) (string, error)
+	Errors() []npgdiff.DiffError
 }
 
-// Command defines the netpol connectivity diff command tree
-func Command(cliEnvironment environment.Environment) *cobra.Command {
-	diffNetpolCmd := &diffNetpolCommand{env: cliEnvironment}
-	c := &cobra.Command{
-		Use:   "diff",
-		Short: "(Technology Preview) Report connectivity-diff based on two directories containing network policies and YAML manifests with workload resources.",
-		Long:  `Based on two input folders containing Kubernetes workloads and network policy YAMLs, this command will report all differences in allowed connections between the resources.` + common.TechPreviewLongText,
-
-		Args: cobra.ExactArgs(0),
-		RunE: func(c *cobra.Command, args []string) error {
-			diffNetpolCmd.env.Logger().WarnfLn("This is a Technology Preview feature. Red Hat does not recommend using Technology Preview features in production.")
-			analyzer, err := diffNetpolCmd.construct()
-			if err != nil {
-				return err
-			}
-			if err := diffNetpolCmd.validate(); err != nil {
-				return err
-			}
-			warns, errs := diffNetpolCmd.analyzeConnectivityDiff(analyzer)
-			err = resources.SummarizeErrors(warns, errs, diffNetpolCmd.treatWarningsAsErrors, diffNetpolCmd.env.Logger())
-			if err != nil {
-				return errors.Wrap(err, "running command")
-			}
-			return nil
-		},
-	}
-
-	c.Flags().StringVarP(&diffNetpolCmd.inputFolderPath1, "dir1", "", "", "first dir path of input resources (required)")
-	c.Flags().StringVarP(&diffNetpolCmd.inputFolderPath2, "dir2", "", "", "second dir path of input resources to be compared with the first dir path (required)")
-	c.Flags().BoolVar(&diffNetpolCmd.treatWarningsAsErrors, "strict", false, "treat warnings as errors")
-	c.Flags().BoolVar(&diffNetpolCmd.stopOnFirstError, "fail", false, "fail on the first encountered error")
-	c.Flags().BoolVar(&diffNetpolCmd.removeOutputPath, "remove", false, "remove the output path if it already exists")
-	c.Flags().BoolVar(&diffNetpolCmd.outputToFile, "save-to-file", false, "whether to save connections diff output into default file")
-	c.Flags().StringVarP(&diffNetpolCmd.outputFilePath, "output-file", "f", "", "save connections diff output into specific file")
-	c.Flags().StringVarP(&diffNetpolCmd.outputFormat, "output-format", "o", defaultOutputFormat, "configure the connections diff in specific format, supported formats: txt|md|csv")
-	return c
+func (cmd *diffNetpolCommand) processInput() (info1 []*resource.Info, info2 []*resource.Info, warnings []error, errs []error) {
+	infos1, warns1, errs1 := resources.GetK8sInfos(cmd.inputFolderPath1, cmd.stopOnFirstError, cmd.treatWarningsAsErrors)
+	infos2, warns2, errs2 := resources.GetK8sInfos(cmd.inputFolderPath2, cmd.stopOnFirstError, cmd.treatWarningsAsErrors)
+	return infos1, infos2, append(warns1, warns2...), append(errs1, errs2...)
 }
 
-func (cmd *diffNetpolCommand) construct() (diffAnalyzer, error) {
-	var opts []npguard.DiffAnalyzerOption
-	if cmd.env != nil && cmd.env.Logger() != nil {
-		opts = append(opts, npguard.WithLogger(npg.NewLogger(cmd.env.Logger())))
+func (cmd *diffNetpolCommand) analyzeConnectivityDiff(analyzer diffAnalyzer) (w []error, e []error) {
+	info1, info2, warns, errs := cmd.processInput()
+	if cmd.stopOnFirstError && (len(errs) > 0 || (len(warns) > 0 && cmd.treatWarningsAsErrors)) {
+		return warns, errs
 	}
-	if cmd.stopOnFirstError {
-		opts = append(opts, npguard.WithStopOnError())
+
+	connsDiff, err := analyzer.ConnDiffFromResourceInfos(info1, info2)
+	if err != nil {
+		if errors.Is(err, os.ErrNotExist) {
+			return warns, append(errs, errox.NotFound.Newf(err.Error()))
+		}
+		return warns, append(errs, errors.Wrap(err, "connectivity diff analysis"))
 	}
-	if cmd.outputFormat != "" {
-		opts = append(opts, npguard.WithOutputFormat(cmd.outputFormat))
+	connsDiffStr, err := analyzer.ConnectivityDiffToString(connsDiff)
+	if err != nil {
+		return warns, append(errs, errors.Wrap(err, "formatting connectivity diff"))
 	}
-	opts = append(opts, npguard.WithArgNames("dir1", "dir2"))
-	if cmd.outputFilePath != "" {
-		cmd.outputToFile = true
+	if err := cmd.outputConnsDiff(connsDiffStr); err != nil {
+		return warns, append(errs, err)
 	}
-	return npguard.NewDiffAnalyzer(opts...), nil
+
+	w, e = npg.HandleNPGuardErrors(analyzer.Errors())
+	return append(warns, w...), append(errs, e...)
 }
 
-func (cmd *diffNetpolCommand) validate() error {
-	if cmd.inputFolderPath1 == "" {
-		return errox.InvalidArgs.Newf("--dir1 is required")
+func (cmd *diffNetpolCommand) outputConnsDiff(connsDiffStr string) error {
+	if cmd.outputToFile {
+		if cmd.outputFilePath == "" { // save-to-file is true, but output file path is not provided
+			cmd.outputFilePath = cmd.getDefaultFileName()
+		}
+
+		if err := writeFile(cmd.outputFilePath, connsDiffStr); err != nil {
+			return errors.Wrap(err, "error writing connections diff output")
+		}
 	}
-	if cmd.inputFolderPath2 == "" {
-		return errox.InvalidArgs.Newf("--dir2 is required")
-	}
-	if err := cmd.setupPath(cmd.outputFilePath); err != nil {
-		return errors.Wrap(err, "failed to set up file path")
-	}
+
+	cmd.printConnsDiff(connsDiffStr)
 	return nil
 }
 
-func (cmd *diffNetpolCommand) setupPath(path string) error {
-	if _, err := os.Stat(path); err == nil && !cmd.removeOutputPath {
-		return errox.AlreadyExists.Newf("path %q already exists. Use --remove to overwrite or select a different path.", path)
-	} else if !os.IsNotExist(err) {
-		return errors.Wrapf(err, "failed to check whether path %q exists", path)
+func writeFile(outputPath string, content string) error {
+	if err := os.MkdirAll(filepath.Dir(outputPath), 0755); err != nil {
+		return errors.Wrapf(err, "error creating directory for file %q", outputPath)
 	}
-	return nil
+	return errors.Wrap(os.WriteFile(outputPath, []byte(content), os.FileMode(0644)), "error writing file")
+}
+
+func (cmd *diffNetpolCommand) printConnsDiff(connsDiff string) {
+	cmd.env.Logger().PrintfLn(connsDiff)
+}
+
+func (cmd *diffNetpolCommand) getDefaultFileName() string {
+	fileNamePrefix := defaultOutputFileNamePrefix
+	if cmd.outputFormat == "" {
+		return fileNamePrefix + defaultOutputFormat
+	}
+	return fileNamePrefix + cmd.outputFormat
 }
