+++ roxctl/netpol/connectivity/map/command_test.go
@@ -1,9 +1,7 @@
 package connectivitymap
 
 import (
-	goerrors "errors"
 	"os"
-	"path"
 	"testing"
 
 	"github.com/stackrox/rox/pkg/errox"
@@ -11,16 +9,16 @@ import (
 	"github.com/stretchr/testify/suite"
 )
 
-func TestAnalyzeNetpolCommand(t *testing.T) {
+func TestConnectivityMapCommand(t *testing.T) {
 	t.Parallel()
-	suite.Run(t, new(analyzeNetpolTestSuite))
+	suite.Run(t, new(connectivityMapCommandSuite))
 }
 
-type analyzeNetpolTestSuite struct {
+type connectivityMapCommandSuite struct {
 	suite.Suite
 }
 
-func (d *analyzeNetpolTestSuite) TestValidate() {
+func (d *connectivityMapCommandSuite) TestValidate() {
 	cases := []struct {
 		name               string
 		inputFolderPath    string
@@ -78,215 +76,4 @@ func (d *analyzeNetpolTestSuite) TestValidate() {
 			d.Assert().NoError(err)
 		})
 	}
-}
-
-func (d *analyzeNetpolTestSuite) TestAnalyzeNetpol() {
-	tmpOutFileName := d.T().TempDir() + "/out"
-	outFileTxt := tmpOutFileName + ".txt"
-	outFileJSON := tmpOutFileName + ".json"
-	outFileMD := tmpOutFileName + ".md"
-	outFileCSV := tmpOutFileName + ".csv"
-	outFileDOT := tmpOutFileName + ".dot"
-	cases := map[string]struct {
-		inputFolderPath       string
-		treatWarningsAsErrors bool
-		stopOnFirstErr        bool
-		outFile               string
-		outputToFile          bool
-		focusWorkload         string
-		outputFormat          string
-		removeOutputPath      bool
-
-		expectedErrors   []string
-		expectedWarnings []string
-	}{
-		"Not existing inputFolderPath should print error about path not existing but attempt analysis": {
-			inputFolderPath:       "/tmp/xxx",
-			stopOnFirstErr:        false,
-			treatWarningsAsErrors: false,
-			expectedErrors: []string{
-				"the path \"/tmp/xxx\" does not exist",
-				"no relevant Kubernetes workload resources found"},
-			expectedWarnings: []string{"no relevant Kubernetes network policy resources found"},
-		},
-		"errors with no resources found": {
-			inputFolderPath: "testdata/empty-yamls",
-			expectedErrors:  []string{"no relevant Kubernetes workload resources found"},
-			expectedWarnings: []string{
-				"unable to decode \"testdata/empty-yamls/empty.yaml\": Object 'Kind' is missing in",
-				"unable to decode \"testdata/empty-yamls/empty2.yaml\": Object 'Kind' is missing in",
-				"no relevant Kubernetes network policy resources found",
-			},
-		},
-		"treating warnings as errors": {
-			inputFolderPath:       "testdata/minimal-with-invalid-doc",
-			expectedErrors:        []string{"unable to decode \"testdata/minimal-with-invalid-doc/resources.yaml\""},
-			expectedWarnings:      []string{},
-			treatWarningsAsErrors: true,
-		},
-		"warnings not indicated without strict": {
-			inputFolderPath:  "testdata/minimal-with-invalid-doc",
-			expectedErrors:   []string{},
-			expectedWarnings: []string{"unable to decode \"testdata/minimal-with-invalid-doc/resources.yaml\": Object 'Kind' is missing in"},
-		},
-		"stopOnFistError": {
-			inputFolderPath: "testdata/dirty",
-			stopOnFirstErr:  true,
-			expectedErrors:  []string{"no relevant Kubernetes workload resources found"},
-			expectedWarnings: []string{
-				"error parsing testdata/dirty/backend.yaml: error converting YAML to JSON",
-				"error parsing testdata/dirty/frontend.yaml: error converting YAML to JSON",
-				"no relevant Kubernetes network policy resources found",
-			},
-		},
-		"should override existing file": {
-			inputFolderPath:  "testdata/minimal",
-			outFile:          outFileTxt,
-			removeOutputPath: true,
-			expectedErrors:   []string{},
-			expectedWarnings: []string{"unable to decode \"testdata/minimal/output.json\""},
-		},
-		"output should be written to default txt output file": {
-			inputFolderPath:  "testdata/minimal",
-			outputToFile:     true,
-			outputFormat:     defaultOutputFormat,
-			expectedErrors:   []string{},
-			expectedWarnings: []string{"unable to decode \"testdata/minimal/output.json\""},
-		},
-		"output should be focused to a workload": {
-			inputFolderPath:  "testdata/minimal",
-			focusWorkload:    "default/backend",
-			expectedErrors:   []string{},
-			expectedWarnings: []string{"unable to decode \"testdata/minimal/output.json\""},
-		},
-		"not supported output format": {
-			inputFolderPath:  "testdata/minimal",
-			outputFormat:     "docx",
-			expectedErrors:   []string{"docx output format is not supported."},
-			expectedWarnings: []string{"unable to decode \"testdata/minimal/output.json\""},
-		},
-		"generate output in json format": {
-			inputFolderPath:  "testdata/minimal",
-			outputFormat:     "json",
-			expectedErrors:   []string{},
-			expectedWarnings: []string{"unable to decode \"testdata/minimal/output.json\""},
-
-			outFile: outFileJSON,
-		},
-		"generate output in md format": {
-			inputFolderPath:  "testdata/minimal",
-			outputFormat:     "md",
-			expectedErrors:   []string{},
-			expectedWarnings: []string{"unable to decode \"testdata/minimal/output.json\""},
-
-			outFile: outFileMD,
-		},
-		"generate output in csv format": {
-			inputFolderPath:  "testdata/minimal",
-			outputFormat:     "csv",
-			expectedErrors:   []string{},
-			expectedWarnings: []string{"unable to decode \"testdata/minimal/output.json\""},
-
-			outFile: outFileCSV,
-		},
-		"generate output in dot format": {
-			inputFolderPath:  "testdata/minimal",
-			outputFormat:     "dot",
-			expectedErrors:   []string{},
-			expectedWarnings: []string{"unable to decode \"testdata/minimal/output.json\""},
-
-			outFile: outFileDOT,
-		},
-		"openshift resources are recognized by the serializer with k8s resources": {
-			inputFolderPath: "testdata/frontend-security",
-			expectedErrors:  []string{},
-			expectedWarnings: []string{
-				"Route resource frontend/asset-cache specified workload frontend/asset-cache[Deployment] as a backend, but network policies are blocking ingress connections from an arbitrary in-cluster source to this workload. Connectivity map will not include a possibly allowed connection between the ingress controller and this workload.",
-				"Route resource frontend/webapp specified workload frontend/webapp[Deployment] as a backend, but network policies are blocking ingress connections from an arbitrary in-cluster source to this workload. Connectivity map will not include a possibly allowed connection between the ingress controller and this workload.",
-			},
-		},
-		"output should be written to default json output file": {
-			inputFolderPath:  "testdata/minimal",
-			expectedErrors:   []string{},
-			expectedWarnings: []string{"unable to decode \"testdata/minimal/output.json\""},
-
-			outputToFile: true,
-			outputFormat: "json",
-		},
-		"generate connections list with ingress controller": {
-			inputFolderPath:  "testdata/acs-security-demos",
-			expectedErrors:   []string{},
-			expectedWarnings: []string{},
-			outputToFile:     true,
-		},
-	}
-
-	for name, tt := range cases {
-		tt := tt
-		d.Run(name, func() {
-			env, _, _ := mocks.NewEnvWithConn(nil, d.T())
-			analyzeNetpolCmd := Cmd{
-				stopOnFirstError:      tt.stopOnFirstErr,
-				treatWarningsAsErrors: tt.treatWarningsAsErrors,
-				inputFolderPath:       "", // set through construct
-				outputFilePath:        tt.outFile,
-				removeOutputPath:      tt.removeOutputPath,
-				outputToFile:          tt.outputToFile,
-				focusWorkload:         tt.focusWorkload,
-				outputFormat:          tt.outputFormat,
-				env:                   env,
-			}
-
-			analyzer, err := analyzeNetpolCmd.construct([]string{tt.inputFolderPath})
-			d.Assert().NoError(err)
-			d.Assert().NoError(analyzeNetpolCmd.validate())
-			warns, errs := analyzeNetpolCmd.analyze(analyzer)
-			d.Require().Lenf(errs, len(tt.expectedErrors), "number of errors should be %d", len(tt.expectedErrors))
-			d.Require().Lenf(warns, len(tt.expectedWarnings), "number of warnings should be %d", len(tt.expectedWarnings))
-
-			for _, expError := range tt.expectedErrors {
-				if expError != "" {
-					d.Require().Error(goerrors.Join(errs...))
-					d.Assert().ErrorContainsf(goerrors.Join(errs...), expError,
-						"Expected errors to contain %s", tt.expectedErrors)
-				} else {
-					d.Assert().NoError(goerrors.Join(errs...))
-				}
-			}
-			for _, expWarn := range tt.expectedWarnings {
-				if expWarn != "" {
-					d.Require().Error(goerrors.Join(warns...))
-					d.Assert().ErrorContainsf(goerrors.Join(warns...), expWarn,
-						"Expected warnings to contain %s", tt.expectedWarnings)
-				} else {
-					d.Assert().NoError(goerrors.Join(warns...))
-				}
-			}
-
-			if tt.outFile != "" && len(tt.expectedErrors) == 0 {
-				_, err := os.Stat(tt.outFile)
-				d.Assert().NoError(err) // out file should exist
-			}
-
-			if tt.outputToFile && tt.outFile == "" && len(tt.expectedErrors) == 0 && len(tt.expectedWarnings) == 0 {
-				defaultFile := analyzeNetpolCmd.getDefaultFileName()
-				formatSuffix := ""
-				if tt.outputFormat != "" {
-					d.Assert().Contains(defaultFile, tt.outputFormat)
-					formatSuffix = tt.outputFormat
-				} else {
-					formatSuffix = defaultOutputFormat
-				}
-				output, err := os.ReadFile(defaultFile)
-				d.Assert().NoError(err)
-
-				expectedOutput, err := os.ReadFile(path.Join(tt.inputFolderPath, "output."+formatSuffix))
-				d.Assert().NoError(err)
-				d.Equal(string(expectedOutput), string(output))
-
-				d.Assert().NoError(os.Remove(defaultFile))
-			}
-		})
-	}
-
 }
